requirements.txt
streamlit
   google-generativeai
   requests
   folium
   streamlit-folium
   pillow
